Changed the coordinate arguments to be of type WORD, which is consistent with AmigaOS and the RastPort structure.
git-svn-id: https://svn.aros.org/svn/aros/trunk@12769 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/graphics/move.c
@@ -18,8 +18,8 @@
 
 /*  SYNOPSIS */
 	AROS_LHA(struct RastPort *, rp, A1),
-	AROS_LHA(LONG             , x, D0),
-	AROS_LHA(LONG             , y, D1),
+	AROS_LHA(WORD             , x, D0),
+	AROS_LHA(WORD             , y, D1),
 
 /*  LOCATION */
 	struct GfxBase *, GfxBase, 40, Graphics)
